Enforce ownership and role-based permissions for user and post operations
- Add ownership validation to user and post controllers to restrict operations to resource owners or roles (e.g., ADMIN, MODERATOR).
- Introduce `HttpError` class to standardize error handling and replace redundant throw error functions.
- Extend authentication middleware to allow permit-all access paths.
- Update tests to validate ownership, permission checks, and unauthenticated scenarios.
- Refactor related services and controllers to streamline error handling logic.

+++ src/controllers/post.controller.js
@@ -1,7 +1,11 @@
 import postService from "../services/post.service.js";
+import {HttpError} from "../config/HttpError.js";
 
 class PostController {
     async createPost(req, res, next) {
+        if(req.principal.username !== req.params.author) {
+            throw new HttpError('Not the owner', 403)
+        }
         try {
             const post = await postService.createPost(req.params.author, req.body)
             return res.status(201).json(post)
@@ -20,11 +24,16 @@ class PostController {
     }
 
     async deletePost(req, res, next) {
-        try {
-            const post = await postService.deletePost(req.params.id)
-            return res.json(post)
-        } catch (error) {
-            return next(error)
+        const post = await postService.getPostById(req.params.id)
+        if (req.principal.username === post.author || req.principal.roles.includes('MODERATOR')) {
+            try {
+                const postDeleted = await postService.deletePost(req.params.id)
+                return res.json(postDeleted)
+            } catch (error) {
+                return next(error)
+            }
+        } else {
+            throw new HttpError('Not the owner', 403)
         }
     }
 
@@ -47,6 +56,9 @@ class PostController {
     }
 
     async addComment(req, res, next) {
+        if(req.principal.username !== req.params.commenter) {
+            throw new HttpError('Not the owner', 403)
+        }
         try {
             const comment = await postService.addComment(req.params.id, req.params.commenter, req.body.message)
             return res.json(comment)
@@ -74,6 +86,10 @@ class PostController {
     }
 
     async updatePost(req, res, next) {
+        const post = await postService.getPostById(req.params.id)
+        if (req.principal.username !== post.author) {
+            throw new HttpError('Not the owner', 403)
+        }
         try {
             const postUpdated = await postService.updatePost(req.params.id, req.body)
             return res.json(postUpdated)
