+++ src/controllers/userAccount.controller.js
@@ -1,4 +1,5 @@
 import userAccountService from "../services/userAccount.service.js";
+import {HttpError} from "../config/HttpError.js";
 
 class UserAccountController {
     async register(req, res, next) {
@@ -11,14 +12,21 @@ class UserAccountController {
     }
 
     async deleteUser(req, res, next) {
-        try {
-            const user = await userAccountService.deleteUser(req.params.login);
-            return res.json(user);
-        } catch (error) {
-            return next(error)
+        if(req.principal.username === req.params.login || req.principal.roles.includes('ADMIN')) {
+            try {
+                const user = await userAccountService.deleteUser(req.params.login);
+                return res.json(user);
+            } catch (error) {
+                return next(error)
+            }
+        } else {
+            throw new HttpError('Only owner or admin can delete this user', 403)
         }
     }
     async updateUser(req, res, next) {
+        if(req.principal.username !== req.params.login) {
+            throw new HttpError('Not the owner', 403)
+        }
         try {
             const user = await userAccountService.updateUser(req.params.login, req.body);
             return res.json(user);
@@ -27,6 +35,9 @@ class UserAccountController {
         }
     }
     async addRole(req, res, next) {
+        if(!req.principal.roles.includes('ADMIN')){
+            throw new HttpError('Only admin can change roles', 403)
+        }
         try {
             const user = await userAccountService.changeRole(req.params.login, req.params.role, true);
             return res.json(user);
@@ -36,6 +47,9 @@ class UserAccountController {
     }
 
     async deleteRole(req, res, next) {
+        if(!req.principal.roles.includes('ADMIN')){
+            throw new HttpError('Only admin can change roles', 403)
+        }
         try {
             const user = await userAccountService.changeRole(req.params.login, req.params.role, false);
             return res.json(user);
